Wait until display is used to try to load matplotlib
since display is imported automatically when ruptures is imported,
any use of ruptures in an environment where matplotlib is available
creates an annoying icon in the tasks manager even when ruptures
is only used non-interactively (at least on macOs, not tested on
other platforms).

As a side benefit, such non-interactive use of ruptures will probably
load faster.